+++ services/web-app/pages/assets/templates.js
@@ -9,15 +9,15 @@ import { useContext, useEffect } from 'react'
 
 import AssetsCatalog from '@/components/assets-catalog'
 import PageHeader from '@/components/page-header'
+import { assetTypes } from '@/data/asset-types'
 import { assetsNavData } from '@/data/nav-data'
-import { type } from '@/data/type'
 import { LayoutContext } from '@/layouts/layout'
 import { getAllLibraries } from '@/lib/github'
 
 const Templates = ({ librariesData }) => {
   const { setPrimaryNavData } = useContext(LayoutContext)
 
-  const { template } = type
+  const { template } = assetTypes
 
   const seo = {
     title: 'Templates'